Stop using russian doll cache for context menu
We only used a specific cache on the context menu to ensure changes to
plugins would invalidate the nested cache item and update the menu. This
failed us in a couple of ways because the menu contains multilpe items:
 - Plugins based on instance
 - Links to Send To items based on project
 - Access to the links based on user permissions

This means the cache would become so user specific it defeats it's own
purpose. Instead, we can simply move the wider "page content" cache down
the tree and skip caching the context menu.